refactor(markdown): split the leaf half out of customInlineTransform
Making the walk iterative took customInlineTransform.ts to 302 lines, two over
the limit. The gate offers a baseline entry; that is not the fix.

customInlineMarks.ts takes the leaf half — what a node IS (isTextNode,
isSkippableNode) and turning the markers inside one text value into mark nodes
(parseMarksInText, createMarkNode), with SKIP_NODE_TYPES and MarkName.

That seam rather than another because the dependency runs ONE way: nothing in
the new file reaches back into the walk, so the walk and the cross-sibling span
logic can both import it without a cycle. Splitting the span logic instead
would have needed these same predicates and pointed the arrow back.

    customInlineTransform.ts  302 -> 229
    customInlineMarks.ts      new, 108

Verified: 5,435 markdown-pipeline tests, typecheck, lint:file-size,
lint:test-types, lint:header-refs, knip and eslint all clean.
